Move most operations to the public JxlColorEncoding
They are safer and sometimes more convenient to write since chromaticity
data and gamma values are colocated with the corresponding “custom” enum
entries and accessible only if it makes sense to do so.
